Simplify integration tests base classes hierarchy
