+++ src/test/groovy/hudson/plugins/gradle/BaseMavenIntegrationTest.groovy
@@ -1,6 +1,5 @@
-package hudson.plugins.gradle.injection
+package hudson.plugins.gradle
 
-import hudson.plugins.gradle.AbstractIntegrationTest
 import org.junit.Rule
 import org.junit.rules.RuleChain
 
@@ -8,6 +7,8 @@ import org.junit.rules.RuleChain
  * Base class for tests that need a Jenkins instance and Maven tool.
  */
 abstract class BaseMavenIntegrationTest extends AbstractIntegrationTest {
+
+    public final MavenInstallationRule mavenInstallationRule = new MavenInstallationRule(j)
 
     @Rule
     public final RuleChain rules = RuleChain.outerRule(noSpaceInTmpDirs).around(j).around(mavenInstallationRule)
