+++ src/test/groovy/hudson/plugins/gradle/injection/BuildScanInjectionMavenCrossVersionTest.groovy
@@ -1,14 +1,16 @@
 package hudson.plugins.gradle.injection
 
+import hudson.plugins.gradle.BaseMavenIntegrationTest
 import hudson.plugins.gradle.BuildScanAction
 import hudson.plugins.gradle.BuildScanBuildWrapper
 import hudson.tasks.Maven
 import org.jvnet.hudson.test.CreateFileBuilder
-import org.jvnet.hudson.test.JenkinsRule
 import spock.lang.Unroll
 
 @Unroll
 class BuildScanInjectionMavenCrossVersionTest extends BaseMavenIntegrationTest {
+
+    private static final String MINIMUM_SUPPORTED_MAVEN_VERSION = '3.3.1'
 
     def 'build scan is discovered from Maven build - #mavenVersion'(String mavenVersion) {
         given:
@@ -35,8 +37,7 @@ class BuildScanInjectionMavenCrossVersionTest extends BaseMavenIntegrationTest {
         def build = j.buildAndAssertSuccess(p)
 
         then:
-        println JenkinsRule.getLog(build)
-        if (mavenVersion >= '3.3.1') {
+        if (mavenVersion >= MINIMUM_SUPPORTED_MAVEN_VERSION) {
             def action = build.getAction(BuildScanAction)
             action.scanUrls.size() == 1
             new URL(action.scanUrls.get(0))
@@ -54,7 +55,8 @@ class BuildScanInjectionMavenCrossVersionTest extends BaseMavenIntegrationTest {
             '3.2.5',
             '3.3.1',
             '3.5.4',
-            '3.8.6'
+            '3.8.7',
+            '3.9.0'
         ]
     }
 }
